+++ openquake/hazardlib/calc/hazard_curve.py
@@ -28,7 +28,7 @@ import numpy
 
 from openquake.baselib.python3compat import raise_, zip
 from openquake.baselib.performance import Monitor
-from openquake.hazardlib.poe import ProbabilityMap, Imtls
+from openquake.hazardlib.probability_map import ProbabilityMap, Imtls
 from openquake.hazardlib.calc import filters
 from openquake.hazardlib.gsim.base import ContextMaker, FarAwayRupture
 from openquake.hazardlib.imt import from_string
@@ -77,21 +77,21 @@ def agg_curves(acc, curves):
     return new
 
 
-def array_of_curves(acc, nsites, imtls, gsim_idx=0):
+def array_of_curves(pmap, nsites, imtls, gsim_idx=0):
     """
-    Convert a dictionary of curves into an array of curves of length `nsites`
+    Convert a probability map into an array of length `nsites`
     and dtype `imt_dt`.
 
-    :param acc: a dictionary sid -> ProbabilityCurve
+    :param pmap: a dictionary sid -> ProbabilityCurve
     :param nsites: the number of sites in the full site collection
     :param imtls: intensity measure types and levels
     :param gsims_idx: extract the data corresponding to a specific GSIM
     """
     curves = numpy.zeros(nsites, imtls.imt_dt)
-    for sid in acc:
-        array = acc[sid].array[:, gsim_idx]
+    for sid in pmap:
+        array = pmap[sid].array[:, gsim_idx]
         for imt in imtls:
-            curves[imt][sid] = array[imtls.slicedic[imt]]
+            curves[sid][imt] = array[imtls.slicedic[imt]]
     return curves
 
 
@@ -202,8 +202,8 @@ def get_probability_no_exceedance(
     return pne_array
 
 
-def probability_map(src, s_sites, imtls, cmaker, trunclevel, bbs,
-                    ctx_mon, pne_mon, disagg_mon):
+def poe_map(src, s_sites, imtls, cmaker, trunclevel, bbs,
+            ctx_mon, pne_mon, disagg_mon):
     """
     Compute the ProbabilityMap generated by the given source. Also,
     store some information in the monitors and optionally in the
@@ -272,9 +272,8 @@ def hazard_curves_per_trt(
     pmap = ProbabilityMap()
     for src, s_sites in source_site_filter(sources_sites):
         t0 = time.time()
-        pmap += probability_map(
-            src, s_sites, imtls, cmaker, truncation_level, bbs,
-            ctx_mon, pne_mon, disagg_mon)
+        pmap += poe_map(src, s_sites, imtls, cmaker, truncation_level, bbs,
+                        ctx_mon, pne_mon, disagg_mon)
         # we are attaching the calculation times to the monitor
         # so that oq-lite (and the engine) can store them
         monitor.calc_times.append((src.id, time.time() - t0))
